@@ -33,7 +33,7 @@ import numpy.typing as npt
 # Define numeric types
 NumberType = TypeVar(
     'NumberType',
-    bound=Union[np.floating, np.integer, np.bool_, float, int, bool],
+    bound=np.floating | np.integer | np.bool_ | float | int | bool,
 )
 NumberType.__doc__ = """Type variable for numeric data types."""
 
@@ -41,23 +41,23 @@ NumberType.__doc__ = """Type variable for numeric data types."""
 # Its definition should be identical to `NumberType`
 _NumberType = TypeVar(  # noqa: PYI018
     '_NumberType',
-    bound=Union[np.floating, np.integer, np.bool_, float, int, bool],
+    bound=np.floating | np.integer | np.bool_ | float | int | bool,
 )
 
 NumpyArray = npt.NDArray[NumberType]
 
-_FiniteNestedList = Union[
-    list[NumberType],
-    list[list[NumberType]],
-    list[list[list[NumberType]]],
-    list[list[list[list[NumberType]]]],
-]
-_FiniteNestedTuple = Union[
-    tuple[NumberType],
-    tuple[tuple[NumberType]],
-    tuple[tuple[tuple[NumberType]]],
-    tuple[tuple[tuple[tuple[NumberType]]]],
-]
+_FiniteNestedList = (
+    list[NumberType]
+    | list[list[NumberType]]
+    | list[list[list[NumberType]]]
+    | list[list[list[list[NumberType]]]]
+)
+_FiniteNestedTuple = (
+    tuple[NumberType]
+    | tuple[tuple[NumberType]]
+    | tuple[tuple[tuple[NumberType]]]
+    | tuple[tuple[tuple[tuple[NumberType]]]]
+)
 
 _ArrayLike1D = Union[
     NumpyArray[NumberType],
