@@ -9,8 +9,8 @@ from typing import TYPE_CHECKING
 from typing import TypeVar
 
 if TYPE_CHECKING:  # pragma: no cover
+    from collections.abc import Callable
     from typing import Any
-    from typing import Callable
 
     from mypy.plugin import ClassDefContext
     from mypy.types import Instance
